P2022-1557 Common validation (not working):
- creates new class Validation.kt
- changes id of EditTexts in Sign In and Sign Up (there were the same id)
