+++ app/src/main/java/com/intive/patronage22/intivi/Fragments/SignInFragment.kt
@@ -6,11 +6,8 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
 import com.intive.patronage22.intivi.MainActivity
-import com.intive.patronage22.intivi.R
 import com.intive.patronage22.intivi.databinding.FragmentSignInBinding
-import java.util.regex.Pattern
 
 class SignInFragment : Fragment() {
 
@@ -21,7 +18,7 @@ class SignInFragment : Fragment() {
         val bind = FragmentSignInBinding.inflate(layoutInflater)
 
         bind.signInButton.setOnClickListener {
-            if (isLoginFormValid()) {
+            if (isValid()) {
                 val intent = Intent(this.requireContext(), MainActivity::class.java)
                 startActivity(intent)
             }
@@ -29,36 +26,5 @@ class SignInFragment : Fragment() {
         return bind.root
     }
 
-    private fun isEmailOK(): Boolean {
-        val editText = requireView().findViewById<EditText>(R.id.loginEditText)
-        val email = editText.text.toString()
-        val localPartLength = email.split("@").first().length
-        val domainPartLength = email.split("@").last().length
-        if (!isEmailValid(email) || localPartLength > 64 || domainPartLength > 255) {
-            editText.error = getString(R.string.emailValidMessage)
-            return false
-        } else {
-            return true
-        }
-    }
-
-    private fun isEmailValid(email: String): Boolean {
-        return Pattern.compile(
-            """(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])"""
-        ).matcher(email).matches()
-    }
-
-    private fun isPasswordOK(): Boolean {
-        val editText = requireView().findViewById<EditText>(R.id.passwordEditText)
-        if (editText.text.length !in (5..40) || (editText.text.toString()
-                .filterNot { it.isWhitespace() }.length != editText.text.length)
-        ) {
-            editText.error = getString(R.string.passwordValidMessage)
-            return false
-        } else {
-            return true
-        }
-    }
-
-    private fun isLoginFormValid() = isEmailOK() and isPasswordOK()
+    private fun isValid() = Validation().isLoginFormValid()
 }
