+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -13,16 +13,16 @@
         tools:layout_editor_absoluteY="-134dp">
 
         <EditText
-            android:id="@+id/loginEditText"
+            android:id="@+id/editTextRegisterEmail"
             android:layout_marginHorizontal="@dimen/loginMarginHorizontal"
             android:layout_marginTop="@dimen/loginEmailMarginTop"
             android:hint = "@string/loginEditTextHint"
-            android:inputType="text"
+            android:inputType="textEmailAddress"
             style="@style/loginEditText"/>
 
         <EditText
-            android:id="@+id/passwordEditText"
-            android:layout_below="@id/loginEditText"
+            android:id="@+id/editTextRegisterPassword"
+            android:layout_below="@id/editTextRegisterEmail"
             android:layout_marginHorizontal="@dimen/loginMarginHorizontal"
             android:layout_marginTop="@dimen/loginPasswordMarginTop"
             android:hint="@string/passwordEditTextHint"
@@ -30,8 +30,8 @@
             style="@style/loginEditText"/>
 
         <EditText
-            android:id="@+id/repeatPassword"
-            android:layout_below="@id/passwordEditText"
+            android:id="@+id/editTextRegisterRepeatPassword"
+            android:layout_below="@id/editTextRegisterPassword"
             android:layout_marginHorizontal="@dimen/loginMarginHorizontal"
             android:layout_marginTop="@dimen/loginPasswordMarginTop"
             android:hint="@string/repeatPassEditTextHint"
